refactor(lyrics): rename lyrics-lrclib.js -> lyrics-cache.js
After the reactive LRCLib fetch path was removed (Phase 6), src/api/
lyrics-lrclib.js is purely a lyrics_cache read + housekeeping module
(getCached / onBoot / purgeOrphans / cacheStats / purgeAll /
purgeTransient); the "lrclib" name was misleading. git mv to
lyrics-cache.js and update the four imports (server.js, api/admin.js,
api/lyrics.js, subsonic/handlers.js) and two filename comments
(schema.js, config.js#getCached). Exported symbols and import aliases are
unchanged. An adversarial grep confirmed these were the only references.

Also fixes a doubly-stale comment in album-art-lib.js: MSTREAM_LRCLIB_BASE
now lives in src/db/lyrics-lookup-lib.js (it moved there with the HTTP
client in Phase 6), not the renamed file — point it at the right home.
